Rename/organize homepage components.
As the "New Policies" components are only used within the homepage, group all
three components into a single folder. This should make finding the pieces
mildly easier.
