+++ ui/components/homepage/new-policies/view.jsx
